requirements.txt
pandas>=1.5.0
numpy>=1.23.0
openpyxl>=3.0.10
scipy>=1.9.0
statsmodels>=0.13.5
linearmodels>=4.28
scikit-learn>=1.1.0
matplotlib>=3.6.0
seaborn>=0.12.0
shap>=0.41.0
kneed>=0.8.3
python-docx>=0.8.11
